+++ R/crest.R
@@ -16,6 +16,11 @@ clearCache <- function () {
     rm(list=ls(all.names=TRUE, envir=cache), envir=cache)
 }
 
+no.cache <- function () {
+    ## Context manager to temporarily turn cache off if it is on
+    temp.option(crest.cache=FALSE)
+}
+
 ## deal with query params?
 dropCache <- function (x) {
     ## Drop x and anything below it in the tree
@@ -27,7 +32,7 @@ dropOnly <- function (x) {
 }
 dropBelow <- function (x) {
     ## Don't drop x, just those below it in the tree. hence ".+"
-    dropPattern(paste0("^", regexEscape(popQuery(x)), ".+")) 
+    dropPattern(paste0("^", regexEscape(popQuery(x)), ".+"))
 }
 dropPattern <- function (x, escape=TRUE) {
     logMessage("CACHE DROP", x)
@@ -48,7 +53,7 @@ popQuery <- function (x) {
 ##' @importFrom httr GET
 cGET <- function (url, ...) {
     # Always check cache. Just don't write to cache if cache is off
-    
+
     Call <- match.call(expand.dots = TRUE)
     cache.url <- url
     if (!is.null(Call[["query"]])) {
@@ -103,4 +108,3 @@ toQuery <- function (query) {
     }
     return(query)
 }
-