+++ tests/testthat/test-add-subvariable.R
@@ -3,6 +3,7 @@ context('Adding subvariables')
 if (run.integration.tests) {
     with(test.authentication, {
         with(test.dataset(newDatasetFromFixture("apidocs")), {
+<<<<<<< HEAD
             expect_false("doggy daycare" %in% names(subvariables(ds$petloc)))
             ds$petloc_daycare <- VariableDefinition(factor(rep(c("Cat", "Dog"), 10)), name="doggy daycare")
             expect_true('petloc_daycare' %in% aliases(variables(ds)))
@@ -20,6 +21,19 @@ if (run.integration.tests) {
             print(names(subvariables(ds$petloc)))
             expect_true("doggy daycare" %in% names(subvariables(ds$petloc)))
             expect_identical(c("Home", "Work", "doggy daycare"), names(subvariables(ds$petloc)))
+=======
+            test_that("Adding a subvariable to an array", {
+                ds$petloc_daycare <- VariableDefinition(factor(rep(c("Cat",
+                    "Dog"), 10)), name="doggy daycare")
+                expect_identical(c("Home", "Work"),
+                    names(subvariables(ds$petloc)))
+                expect_true("petloc_daycare" %in% aliases(variables(ds)))
+                addSubvariable(ds$petloc, ds$petloc_daycare)
+                ds <- refresh(ds)
+                expect_identical(c("Home", "Work", "doggy daycare"),
+                    names(subvariables(ds$petloc)))
+            })
+>>>>>>> upstream/master
         })
     })
 }