+++ NEWS.md
@@ -1,4 +1,10 @@
-### crunch 1.7.9
+### crunch 1.7.12
+* Add method for R logical &/| Crunch expression
+* Upgrade for compatibility with httr 1.1
+
+### crunch 1.7.10
+* `addSubvariable` function to add to array and multiple response variables (#7)
+* Make paginated requests to GET /values/ for greater reliability
 
 ### crunch 1.7.8
 * Update to match changes in filter API